backend2/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
distlib==0.3.5
dj-database-url==0.5.0
Django==3.2
django-extensions==3.2.0
django-localflavor==3.1
django-reversion==5.0.2
djangorestframework==3.14.0
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.7.1
identify==2.5.3
idna==3.3
localflavor==1.9
nodeenv==1.7.0
numpy==1.23.2
pipreqs==0.4.11
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
python-stdnum==1.17
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.11
virtualenv==20.16.3
yarg==0.1.9
